refactor: unify score viewer prototype
